refactor(md-ast): extract zfb-md-ast crate with shared traits + HastNode
Moves MdastVisitor, HastVisitor traits + HastNode enum from zfb-content
into a new tiny workspace crate crates/zfb-md-ast/. zfb-content re-exports
the moved types via `pub use zfb_md_ast::{...}` so existing public API
stays compatible. Breaks the circular-dep risk between zfb-content and
the future zfb-md-extras crate.

Pure code move with re-export — no behavior change.

Implements #565 (epic #564, Wave 1).
